CPLEX files
The ProjectOAJan zip folder contains all the CPLEX files (i.e., .mod and .dat) used in this study. 
ProjectOA.mod stands for the OPL model of the scenario where no existing stations were considered, whereas ProjectOAExisting.mod refers to the OPL model when forcing the existing stations into the solution. The DataStations.dat file describes the connection with the matiz_distancia.xlsx file that contains the parameter values needed to solve the problem (i.e., the set of stations and the matrix of the shortest distance).